# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_alemancenter_fiber-api_internal_utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      pagination:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      success:
        type: boolean
      token:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_utils.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
host: api.alemancenter.com
info:
  contact:
    email: support@alemancenter.com
    name: API Support
  description: Backend API for Alemancenter Educational Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Alemancenter API
  version: 2.0.0
paths:
...
  /keywords:
    get:
      description: Returns a paginated list of keywords used in articles and/or posts
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      - description: 'Type: all, article, post'
        in: query
        name: type
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      summary: List Keywords
      tags:
      - Keywords
  /keywords/{keyword}:
    get:
      description: Returns a list of articles and posts associated with a specific
        keyword
      parameters:
      - description: Country ID
        in: header
        name: X-Country-Id
        type: string
      - description: The exact keyword text
        in: path
        name: keyword
        required: true
        type: string
      - description: Search query inside keyword results
        in: query
        name: q
        type: string
      - description: Sort order (latest, popular)
        in: query
        name: sort
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      summary: Get Keyword Content
      tags:
      - Keywords
...
schemes:
- https
- http
security:
- FrontendKeyAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  FrontendKeyAuth:
    description: Frontend identifier key for public endpoints.
    in: header
    name: X-Frontend-Key
    type: apiKey
swagger: "2.0"